Feat(be): Spring REST docs 적용 (#26)
* Chore(be): 패키지 구조 변경

* Chore(be): spring rest docs gradle 설정

* Feat(be): adoc 파일 작성

* Feat(be): restdocs config 작성 및 review test 연동

* Chore(be): gitignore에 docs html 추가

* Refactor(be): 일부 리뷰 API 인증 기능 통합

* Chore(be): adoc에 새 테스트 추가

* Test(be): `LoginController` 테스트 구현

* Docs(be): 로그인 API docs 작성

* Fix(be): 리뷰 수정/삭제 로직에서 NPE 발생 문제 해결

* Fix(be): 유저가 작성한 리뷰 가져오기 로직 수정

일대다 관계 수정 불가 이슈 해결

* Fix(be): 테스트 로직 변경

+++ backend/grabtable/src/main/java/edu/skku/grabtable/review/service/ReviewService.java
@@ -1,19 +1,20 @@
-package edu.skku.grabtable.service;
+package edu.skku.grabtable.review.service;
 
 import edu.skku.grabtable.common.exception.BadRequestException;
 import edu.skku.grabtable.common.exception.ExceptionCode;
-import edu.skku.grabtable.domain.Review;
 import edu.skku.grabtable.domain.Store;
 import edu.skku.grabtable.domain.User;
-import edu.skku.grabtable.domain.request.ReviewRequest;
-import edu.skku.grabtable.domain.response.ReviewResponse;
-import edu.skku.grabtable.domain.response.ReviewSummaryResponse;
-import edu.skku.grabtable.repository.ReviewRepository;
 import edu.skku.grabtable.repository.StoreRepository;
 import edu.skku.grabtable.repository.UserRepository;
+import edu.skku.grabtable.review.domain.Review;
+import edu.skku.grabtable.review.domain.request.ReviewRequest;
+import edu.skku.grabtable.review.domain.response.ReviewResponse;
+import edu.skku.grabtable.review.domain.response.ReviewSummaryResponse;
+import edu.skku.grabtable.review.repository.ReviewRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -31,10 +32,12 @@ public class ReviewService {
 
     @Transactional(readOnly = true)
     public List<ReviewResponse> getAllReviewsByUser(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new BadRequestException(ExceptionCode.NOT_FOUND_USER_ID));
+//        User user = userRepository.findById(userId)
+//                .orElseThrow(() -> new BadRequestException(ExceptionCode.NOT_FOUND_USER_ID));
 
-        return user.getReviews().stream().map(ReviewResponse::of).toList();
+        List<Review> reviews = reviewRepository.findByUserId(userId);
+
+        return reviews.stream().map(ReviewResponse::of).toList();
     }
 
     @Transactional(readOnly = true)
@@ -45,22 +48,32 @@ public class ReviewService {
         return store.getReviews().stream().map(ReviewResponse::of).toList();
     }
 
-    public void delete(Long reviewId) {
+    public void delete(Long userId, Long reviewId) {
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new BadRequestException(ExceptionCode.NOT_FOUND_REVIEW_ID));
+
+        User user = review.getUser();
+        if (user == null || !Objects.equals(user.getId(), userId)) {
+            throw new BadRequestException(ExceptionCode.INVALID_REQUEST);
+        }
+
         reviewRepository.delete(review);
     }
 
-    public void update(Long reviewId, String message, Double rating) {
+    public void update(Long userId, Long reviewId, String message, Double rating) {
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new BadRequestException(ExceptionCode.NOT_FOUND_STORE_ID));
+        User user = review.getUser();
+        if (user == null || !Objects.equals(user.getId(), userId)) {
+            throw new BadRequestException(ExceptionCode.INVALID_REQUEST);
+        }
         review.update(message, rating);
     }
 
-    public void upload(ReviewRequest request) {
+    public void upload(Long userId, ReviewRequest request) {
         Store store = storeRepository.findById(request.getStoreId())
                 .orElseThrow(() -> new BadRequestException(ExceptionCode.NOT_FOUND_STORE_ID));
-        User user = userRepository.findById(request.getUserId())
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new BadRequestException(ExceptionCode.NOT_FOUND_USER_ID));
 
         Review review = Review.of(store, user, request.getMessage(), request.getRating());
